fix: dont panic when unable to obtain hostname
if we are unable to obtain the hostname of the local machine from the
/etc/hostname file or HOSTNAME env var then an empty string is used as
the hostname and a warning is shown instead of a panic. since the
hostname is just used to determine if we are executing on the cluster,
and the cluster machines all have their hostnames set, then using the
empty string should not be a problem.