fix: improve error message when job has 0 machines
when a job is not yet in running state the program runs but the function
that lists the job's machines returns zero machines which can cause a
panic from division by 0 which is not a very helpful message. this
commit improves the error/warn messages when no machines are listed for
a job.